fix: address code review findings from CodeRabbit
Fix critical bugs: ServerFeatures.Count JsonPropertyName inside XML comment, ChannelAuthorize UInt32 overflow, ServerInfo duration_us parsed as milliseconds, CloseTimeIso missing converter, CredentialQuery wire field name. Fix validation: Amount XRP null-check, AccountId null handling, Uint64 hex length, OracleConverters truncation/negative/hex-sniffing, LOOracle missing hex converters, BinaryCodec null checks. Refactor: ISerializedType.ToJson object to JsonNode, LedgerObjectConverter deduplicate switch, ScientificDecimalConverter precision, EnumMemberValueConverter strict parsing. Add CHANGES.md entry for v10.3.

+++ Xrpl/Models/Methods/ChannelAuthorize.cs
@@ -87,7 +87,7 @@ namespace Xrpl.Models.Methods
         public double RippleAmount
         {
             get => (double) Amount / 1000000;
-            set => Amount =  Convert.ToUInt32(value * 1000000);
+            set => Amount = Convert.ToUInt64(value * 1000000);
         }
     }
 }
